+++ CHANGELOG.md
@@ -50,6 +50,8 @@ and this project adheres to
   - [#1619](https://github.com/iovisor/bpftrace/issues/1619)
 - Array improvements (support assignment to variables and usage as a map key)
   - [#1656](https://github.com/iovisor/bpftrace/pull/1656)
+- Add builtin function: `macaddr`
+  - [#1647](https://github.com/iovisor/bpftrace/pull/1647)
 
 #### Changed
 - Warn if using `print` on `stats` maps with top and div arguments